+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/store/GoogleFitVariable.kt
@@ -6,5 +6,6 @@ import com.google.android.gms.fitness.data.Field
 data class GoogleFitVariable (
     val dataType : DataType,
     val fields : List<Field>,
-    var allowedOperations : List<String>
+    var allowedOperations : List<String>,
+    val filterBySession : List<String> = arrayListOf()
 )