+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/store/HealthFitnessManager.kt
@@ -2,28 +2,21 @@ package com.outsystems.plugins.healthfitnesslib.store
 
 import android.app.Activity
 import android.app.PendingIntent
+import android.content.Context
+import android.content.Intent
+import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.fitness.Fitness
 import com.google.android.gms.fitness.FitnessOptions
-import android.content.Context
-import android.util.Log
-import androidx.room.Room
-import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.fitness.data.DataSet
 import com.google.android.gms.fitness.data.DataSource
 import com.google.android.gms.fitness.request.DataUpdateListenerRegistrationRequest
 import com.google.android.gms.fitness.request.SensorRequest
 import com.google.android.gms.fitness.result.DataReadResponse
+import com.google.android.gms.fitness.result.SessionReadResponse
 import com.outsystems.plugins.healthfitness.HealthFitnessError
 import com.outsystems.plugins.healthfitnesslib.background.BackgroundJobParameters
-import com.outsystems.plugins.healthfitnesslib.background.DatabaseManager
-import com.outsystems.plugins.healthfitnesslib.background.database.AppDatabase
-import com.outsystems.plugins.healthfitnesslib.background.database.BackgroundJob
-import com.outsystems.plugins.healthfitnesslib.background.database.Notification
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
-import java.lang.Exception
+import com.outsystems.plugins.healthfitnesslib.background.VariableUpdateService
 import java.util.concurrent.TimeUnit
 
 class HealthFitnessManager(var context : Context, var activity : Activity? = null): HealthFitnessManagerInterface {
@@ -91,10 +84,10 @@ class HealthFitnessManager(var context : Context, var activity : Activity? = nul
     }
 
     override fun subscribeToSensorUpdates(variable: GoogleFitVariable,
+                                          variableName : String,
                                           grouping: Long,
                                           jobFrequency: TimeUnit,
                                           parameters: BackgroundJobParameters,
-                                          pendingIntent : PendingIntent,
                                           onSuccess: () -> Unit,
                                           onFailure: (Exception) -> Unit) {
 
@@ -110,7 +103,7 @@ class HealthFitnessManager(var context : Context, var activity : Activity? = nul
                     .setDataType(variable.dataType)
                     .setSamplingRate(grouping, jobFrequency)
                     .build(),
-                pendingIntent
+                getSubscritionPendingIntent(variableName)
             )
             .addOnSuccessListener {
                 onSuccess()
@@ -121,7 +114,7 @@ class HealthFitnessManager(var context : Context, var activity : Activity? = nul
     }
 
     override fun subscribeToHistoryUpdates(variable: GoogleFitVariable,
-                                           pendingIntent : PendingIntent,
+                                           variableName : String,
                                            onSuccess: () -> Unit,
                                            onFailure: (Exception) -> Unit) {
 
@@ -133,7 +126,7 @@ class HealthFitnessManager(var context : Context, var activity : Activity? = nul
 
         val request = DataUpdateListenerRegistrationRequest.Builder()
             .setDataType(variable.dataType)
-            .setPendingIntent(pendingIntent)
+            .setPendingIntent(getSubscritionPendingIntent(variableName))
             .build()
 
         Fitness.getHistoryClient(context, account)
@@ -158,4 +151,11 @@ class HealthFitnessManager(var context : Context, var activity : Activity? = nul
         return GoogleSignIn.getLastSignedInAccount(context)
     }
 
+    private fun getSubscritionPendingIntent(variableName : String) : PendingIntent {
+        //do the actual subscription to the variable updates
+        val intent = Intent(context, VariableUpdateService::class.java)
+        intent.putExtra(VariableUpdateService.VARIABLE_NAME, variableName)
+        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+        return PendingIntent.getBroadcast(context, variableName.hashCode(), intent, PendingIntent.FLAG_UPDATE_CURRENT)
+    }
 }