+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/store/HealthFitnessManagerInterface.kt
@@ -4,6 +4,7 @@ import android.app.PendingIntent
 import com.google.android.gms.fitness.FitnessOptions
 import com.google.android.gms.fitness.data.DataSet
 import com.google.android.gms.fitness.result.DataReadResponse
+import com.google.android.gms.fitness.result.SessionReadResponse
 import com.outsystems.plugins.healthfitnesslib.background.BackgroundJobParameters
 import java.lang.Exception
 import java.util.concurrent.TimeUnit
@@ -14,6 +15,7 @@ interface HealthFitnessManagerInterface {
     fun areGoogleFitPermissionsGranted(options: FitnessOptions?): Boolean
     fun requestPermissions(fitnessOptions: FitnessOptions, resultCode: Int)
     fun updateDataOnStore(dataSet: DataSet?, onSuccess: () -> Unit, onFailure: (Exception) -> Unit)
+
     fun getDataFromStore(queryInformation: AdvancedQuery,
                          onSuccess: (DataReadResponse) -> Unit,
                          onFailure: (Exception) -> Unit)
@@ -24,15 +26,15 @@ interface HealthFitnessManagerInterface {
                                     onFailure: (Exception) -> Unit)
 
     fun subscribeToSensorUpdates(variable: GoogleFitVariable,
+                                 variableName: String,
                                  grouping: Long,
                                  jobFrequency: TimeUnit,
                                  parameters: BackgroundJobParameters,
-                                 pendingIntent: PendingIntent,
                                  onSuccess: () -> Unit,
                                  onFailure: (Exception) -> Unit)
 
     fun subscribeToHistoryUpdates(variable: GoogleFitVariable,
-                                  pendingIntent : PendingIntent,
+                                  variableName: String,
                                   onSuccess: () -> Unit,
                                   onFailure: (Exception) -> Unit)
 }