+++ CHANGELOG.md
@@ -8,11 +8,15 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
-## 2021-11-12
+## 2021-11-18
 - Implementation of new variables and new version of background job iOS (https://outsystemsrd.atlassian.net/browse/RMET-1138)
 
 ## 2021-11-12
 - Implementation of the background job iOS (https://outsystemsrd.atlassian.net/browse/RMET-1133)
+
+- Test: Added setBackgroundJob unit tests for Android (https://outsystemsrd.atlassian.net/browse/RMET-1191)
+
+- Implementation of the background job (https://outsystemsrd.atlassian.net/browse/RMET-1133)
 
 ## 2021-11-11
 - Implemented SetBackgroundJob action with notifications (https://outsystemsrd.atlassian.net/browse/RMET-1070)
@@ -42,7 +46,7 @@ The changes documented here do not include those from the original repository.
 
 ## [Version 1.0.0]
 
-## 2021-10-04<>
+## 2021-10-04
 - Refactorings for Android (https://outsystemsrd.atlassian.net/browse/RMET-1066)
 
 ## 2021-09-21